add tetgen and lots of test

+++ include/ArmorerPhys/sim/pd/pd_solver_3d.cpp
@@ -153,8 +153,8 @@ void ProjectiveDynamicsSolver3D::localStep(const MatxXd& verts,
 void ProjectiveDynamicsSolver3D::globalStep(MatxXd& verts,
                                             const MatxXd& verts_pred) {
   Eigen::MatrixXd rhs = M_h2 * verts_pred + J * P;
-  Eigen::MatrixXd result = sparse_solver.solve(rhs.cast<double>());
-  verts = result.cast<double>();
+  Eigen::MatrixXd result = sparse_solver.solve(rhs);
+  verts = result;
 }
 
 }  // namespace aphys
